Reorganizing map sidebar files - location selector now shared by different panels
