+++ frontend/src/containers/map/sidebar/main-panel/location-selector/location-dropdown/index.tsx
