+++ frontend/src/containers/map/sidebar/main-panel/location-selector/type-toggle/index.tsx
