+++ frontend/src/containers/map/sidebar/main-panel/panels/details/countries-list/index.tsx
