+++ frontend/src/containers/map/sidebar/main-panel/panels/details/details-button/index.tsx
