+++ frontend/src/containers/map/sidebar/main-panel/panels/details/widgets/establishment-stages/index.tsx
