+++ frontend/src/containers/map/sidebar/main-panel/panels/details/widgets/fishing-protection/index.tsx
