+++ frontend/src/containers/map/sidebar/main-panel/panels/details/widgets/index.tsx
