+++ frontend/src/containers/map/sidebar/main-panel/panels/details/widgets/marine-conservation/index.tsx
