+++ frontend/src/containers/map/sidebar/main-panel/panels/details/widgets/protection-types/index.tsx
