+++ frontend/src/containers/map/sidebar/main-panel/panels/index.tsx
@@ -1,21 +1,21 @@
 import Details from './details';
 import Modelling from './modelling';
 
-export const SIDEBAR_TYPES = {
+export const PANEL_TYPES = {
   progress_tracker: 'progress-tracker',
   conservation_builder: 'conservation-builder',
 };
 
 const SIDEBAR_COMPONENTS = {
-  [SIDEBAR_TYPES.progress_tracker]: Details,
-  [SIDEBAR_TYPES.conservation_builder]: Modelling,
+  [PANEL_TYPES.progress_tracker]: Details,
+  [PANEL_TYPES.conservation_builder]: Modelling,
 };
 
-type MainPanelProps = {
-  type: keyof typeof SIDEBAR_TYPES;
+type PanelsProps = {
+  type: keyof typeof PANEL_TYPES;
 };
 
-const MainPanel: React.FC<MainPanelProps> = ({ type }) => {
+const MainPanel: React.FC<PanelsProps> = ({ type }) => {
   const Component = SIDEBAR_COMPONENTS[type] || Details;
 
   return <Component />;
